--- src/store/use-pagination.js
@@ -1,47 +1,0 @@
-import { useMemo } from 'react';
-
-export const DOTS = '...';
-
-const range = (start, end) => {
-  let length = end - start + 1;
-  return Array.from({ length }, (_, idx) => idx + start);
-};
-
-export const usePagination = ({ totalCount, pageSize, siblingCount = 1, currentPage }) => {
-  const paginationRange = useMemo(() => {
-    const totalPageCount = Math.ceil(totalCount / pageSize);
-    const totalPageNumbers = siblingCount + 5;
-
-    if (totalPageNumbers >= totalPageCount) {
-      return range(1, totalPageCount);
-    }
-
-    const leftSiblingIndex = Math.max(currentPage - siblingCount, 1);
-    const rightSiblingIndex = Math.min(currentPage + siblingCount, totalPageCount);
-
-    const shouldShowLeftDots = leftSiblingIndex > 2;
-    const shouldShowRightDots = rightSiblingIndex < totalPageCount - 1;
-
-    const firstPageIndex = 1;
-    const lastPageIndex = totalPageCount;
-
-    if (!shouldShowLeftDots && shouldShowRightDots) {
-      let leftItemCount = currentPage < 3 ? 3 : 4;
-      let leftRange = range(1, leftItemCount);
-      return [...leftRange, '...', totalPageCount];
-    }
-
-    if (shouldShowLeftDots && !shouldShowRightDots) {
-      let rightItemCount = currentPage > totalPageCount - 2 ? 3 : 4;
-      let rightRange = range(totalPageCount - rightItemCount + 1, totalPageCount);
-      return [firstPageIndex, '...', ...rightRange];
-    }
-
-    if (shouldShowLeftDots && shouldShowRightDots) {
-      let middleRange = range(leftSiblingIndex, rightSiblingIndex);
-      return [firstPageIndex, `...`, ...middleRange, '...', lastPageIndex];
-    }
-  }, [totalCount, pageSize, siblingCount, currentPage]);
-
-  return paginationRange;
-};
