+++ src/utils.js
@@ -33,3 +33,45 @@ export function codeGenerator(start = 0) {
 export function numberFormat(value, locale = 'ru-RU', options = {}) {
   return new Intl.NumberFormat(locale, options).format(value);
 }
+
+export const DOTS = '...';
+
+const range = (start, end) => {
+  let length = end - start + 1;
+  return Array.from({ length }, (_, idx) => idx + start);
+};
+
+export const pagination = ({ totalCount, pageSize, siblingCount = 1, currentPage }) => {
+  const totalPageCount = Math.ceil(totalCount / pageSize);
+  const totalPageNumbers = siblingCount + 5;
+
+  if (totalPageNumbers >= totalPageCount) {
+    return range(1, totalPageCount);
+  }
+
+  const leftSiblingIndex = Math.max(currentPage - siblingCount, 1);
+  const rightSiblingIndex = Math.min(currentPage + siblingCount, totalPageCount);
+
+  const shouldShowLeftDots = leftSiblingIndex > 2;
+  const shouldShowRightDots = rightSiblingIndex < totalPageCount - 1;
+
+  const firstPageIndex = 1;
+  const lastPageIndex = totalPageCount;
+
+  if (!shouldShowLeftDots && shouldShowRightDots) {
+    let leftItemCount = currentPage < 3 ? 3 : 4;
+    let leftRange = range(1, leftItemCount);
+    return [...leftRange, '...', totalPageCount];
+  }
+
+  if (shouldShowLeftDots && !shouldShowRightDots) {
+    let rightItemCount = currentPage > totalPageCount - 2 ? 3 : 4;
+    let rightRange = range(totalPageCount - rightItemCount + 1, totalPageCount);
+    return [firstPageIndex, '...', ...rightRange];
+  }
+
+  if (shouldShowLeftDots && shouldShowRightDots) {
+    let middleRange = range(leftSiblingIndex, rightSiblingIndex);
+    return [firstPageIndex, `...`, ...middleRange, '...', lastPageIndex];
+  }
+};
